reusable component for image upload button
